feat: module based structure created for local ui
