+++ pebblo/app/pebblo-ui/src/components/card.js
@@ -1,0 +1,5 @@
+export function Card(children) {
+  return `
+      <div class="card overflow-hidden">${children}</div>
+      `;
+}
