+++ pebblo/app/pebblo-ui/src/pages/index.js
@@ -1,0 +1,3 @@
+import { AppDetailsPage } from "./appDetailsPage.js";
+import { OverviewPage } from "./overviewPage.js";
+export { OverviewPage, AppDetailsPage };
