+++ pyproject.toml
@@ -153,7 +153,7 @@ pebblo = "pebblo.app.daemon:start"
 # If there are data files included in your packages that need to be
 # installed, specify them here.
 
-package-data = {"pebblo" = ['reports/templates/*', 'reports/assets/*','app/pebblo-ui/*' ,'app/pebblo-ui/static/*' ]}
+package-data = {"pebblo" = ['reports/templates/*', 'reports/assets/*','app/pebblo-ui/*' ,'app/pebblo-ui/static/*', 'app/pebblo-ui/src/*', 'app/pebblo-ui/src/components/*', 'app/pebblo-ui/src/pages/*'  ]}
 include-package-data = false
 
 [tool.setuptools.packages.find]
